requirements.txt

SpeechRecognition
pyttsx3
pyaudio


nltk
scikit-learn
joblib
numpy


pyautogui
pyperclip


customtkinter
pillow
python-Levenshtein